python/tesart/commutation_matrix/requirements.txt
PyQt5==5.14.1
pyserial
pyvisa-py
netifaces